# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Europass CV Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CvSummary {
  id: number;
  fullName: string;
  /** @nullable */
  email?: string | null;
  mainColor: string;
  createdAt: string;
  updatedAt: string;
}

export interface Experience {
  id: number;
  cvId: number;
  jobTitle: string;
  employer: string;
  /** @nullable */
  city?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder: number;
}

export interface Education {
  id: number;
  cvId: number;
  degree: string;
  institution: string;
  /** @nullable */
  city?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  grade?: string | null;
  sortOrder: number;
}

export interface Language {
  id: number;
  cvId: number;
  languageName: string;
  /** @nullable */
  listening?: string | null;
  /** @nullable */
  reading?: string | null;
  /** @nullable */
  spokenInteraction?: string | null;
  /** @nullable */
  spokenProduction?: string | null;
  /** @nullable */
  writing?: string | null;
  sortOrder: number;
}

export interface Cv {
  id: number;
  fullName: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  idNumber?: string | null;
  /** @nullable */
  passportNumber?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  digitalSkills?: string | null;
  /** @nullable */
  hobbies?: string | null;
  /** @nullable */
  drivingLicense?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  mainColor: string;
  useGraphics: boolean;
  createdAt: string;
  updatedAt: string;
  experiences: Experience[];
  educations: Education[];
  languages: Language[];
}

export interface CvInput {
  /** @minLength 1 */
  fullName: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  idNumber?: string | null;
  /** @nullable */
  passportNumber?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  digitalSkills?: string | null;
  /** @nullable */
  hobbies?: string | null;
  /** @nullable */
  drivingLicense?: string | null;
  mainColor?: string;
  useGraphics?: boolean;
}

export interface ExperienceInput {
  /** @minLength 1 */
  jobTitle: string;
  /** @minLength 1 */
  employer: string;
  /** @nullable */
  city?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
}

export interface EducationInput {
  /** @minLength 1 */
  degree: string;
  /** @minLength 1 */
  institution: string;
  /** @nullable */
  city?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  grade?: string | null;
  sortOrder?: number;
}

export interface LanguageInput {
  /** @minLength 1 */
  languageName: string;
  /** @nullable */
  listening?: string | null;
  /** @nullable */
  reading?: string | null;
  /** @nullable */
  spokenInteraction?: string | null;
  /** @nullable */
  spokenProduction?: string | null;
  /** @nullable */
  writing?: string | null;
  sortOrder?: number;
}