# lib/api-zod/src/generated/types/cvInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Europass CV Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface CvInput {
  /** @minLength 1 */
  fullName: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  idNumber?: string | null;
  /** @nullable */
  passportNumber?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  digitalSkills?: string | null;
  /** @nullable */
  hobbies?: string | null;
  /** @nullable */
  drivingLicense?: string | null;
  mainColor?: string;
  useGraphics?: boolean;
}