# lib/api-zod/src/generated/types/cvSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Europass CV Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface CvSummary {
  id: number;
  fullName: string;
  /** @nullable */
  email?: string | null;
  mainColor: string;
  createdAt: Date;
  updatedAt: Date;
}