# lib/api-zod/src/generated/types/education.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Europass CV Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Education {
  id: number;
  cvId: number;
  degree: string;
  institution: string;
  /** @nullable */
  city?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  grade?: string | null;
  sortOrder: number;
}