# lib/api-zod/src/generated/types/experience.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Europass CV Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Experience {
  id: number;
  cvId: number;
  jobTitle: string;
  employer: string;
  /** @nullable */
  city?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder: number;
}