# lib/api-zod/src/generated/types/experienceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Europass CV Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface ExperienceInput {
  /** @minLength 1 */
  jobTitle: string;
  /** @minLength 1 */
  employer: string;
  /** @nullable */
  city?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
}