# lib/api-zod/src/generated/types/language.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Europass CV Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Language {
  id: number;
  cvId: number;
  languageName: string;
  /** @nullable */
  listening?: string | null;
  /** @nullable */
  reading?: string | null;
  /** @nullable */
  spokenInteraction?: string | null;
  /** @nullable */
  spokenProduction?: string | null;
  /** @nullable */
  writing?: string | null;
  sortOrder: number;
}